fix a build

+++ include/as/as-types.hpp
@@ -23,10 +23,10 @@ using f64 = double;
 // helpers for situations when literals want to be used as a smaller
 // size value without needing to use an old style C cast or verbose
 // C++ cast (e.g. with colors between 0-255)
-inline constexpr u8 operator "" _u8(u64 v) noexcept;
-inline constexpr u16 operator "" _u16(u64 v) noexcept;
-inline constexpr s8 operator "" _s8(u64 v) noexcept;
-inline constexpr s16 operator "" _s16(u64 v) noexcept;
+inline constexpr u8 operator "" _u8(unsigned long long v) noexcept;
+inline constexpr u16 operator "" _u16(unsigned long long v) noexcept;
+inline constexpr s8 operator "" _s8(unsigned long long v) noexcept;
+inline constexpr s16 operator "" _s16(unsigned long long v) noexcept;
 
 // static assert to check library floating point precision has been set correctly
 #if (!defined(AS_PRECISION_FLOAT) && !defined(AS_PRECISION_DOUBLE))
