+++ include/as/as-types.inl
@@ -1,22 +1,22 @@
 namespace as
 {
 
-constexpr u8 operator "" _u8(u64 v) noexcept
+constexpr u8 operator "" _u8(unsigned long long v) noexcept
 {
     return static_cast<u8>(v);
 }
 
-constexpr u16 operator "" _u16(u64 v) noexcept
+constexpr u16 operator "" _u16(unsigned long long v) noexcept
 {
     return static_cast<u16>(v);
 }
 
-constexpr s8 operator "" _s8(u64 v) noexcept
+constexpr s8 operator "" _s8(unsigned long long v) noexcept
 {
     return static_cast<s8>(v);
 }
 
-constexpr s16 operator "" _s16(u64 v) noexcept
+constexpr s16 operator "" _s16(unsigned long long v) noexcept
 {
     return static_cast<s16>(v);
 }
